moving standing wave to main test set
