+++ src/shared/particle_dynamics/general_dynamics/general_interaction.h
@@ -21,17 +21,44 @@
  *                                                                           *
  * ------------------------------------------------------------------------- */
 /**
- * @file    all_general_dynamics.h
- * @brief   This is the header file that user code should include to pick up all
- *          general dynamics used in SPHinXsys.
- * @author	Chi Zhang and Xiangyu Hu
+ * @file 	general_interaction.h
+ * @brief 	This is the interaction dynamics applicable for all type bodies
+ * @author	Yaru Ren and Xiangyu Hu
  */
 
-#pragma once
+#ifndef GENERAL_INTERACTION_H
+#define GENERAL_INTERACTION_H
 
-#include "general_bounding.h"
 #include "general_dynamics.h"
-#include "general_dynamics_refinement.h"
-#include "general_interaction.h"
-#include "general_interpolation.h"
-#include "general_life_time_dynamics.h"
+
+namespace SPH
+{
+class CorrectionMatrixInner : public LocalDynamics, public GeneralDataDelegateInner
+{
+  public:
+    CorrectionMatrixInner(BaseInnerRelation &inner_relation, int beta = 0, Real alpha = Real(0));
+    virtual ~CorrectionMatrixInner(){};
+
+  protected:
+    int beta_;
+    Real alpha_;
+    StdLargeVec<Matd> &B_;
+
+    void interaction(size_t index_i, Real dt = 0.0);
+    void update(size_t index_i, Real dt = 0.0);
+};
+
+class CorrectionMatrixComplex : public CorrectionMatrixInner, public GeneralDataDelegateContact
+{
+  public:
+    CorrectionMatrixComplex(ComplexRelation &complex_relation, int beta = 0, Real alpha = Real(0));
+    virtual ~CorrectionMatrixComplex(){};
+
+  protected:
+    StdVec<StdLargeVec<Real> *> contact_Vol_;
+    StdVec<StdLargeVec<Real> *> contact_mass_;
+
+    void interaction(size_t index_i, Real dt = 0.0);
+};
+} // namespace SPH
+#endif // GENERAL_INTERACTION_H
